app/pyproject.toml
[project]
name = "eventyay-unified"
version = "0.1.0"
description = "Add your description here"
readme = "README.rst"
requires-python = ">=3.11"
keywords = ["tickets", "events", "web", "shop", "ecommerce"]
authors = [ {name = "eventyay", email = "info@eventyay.com"}, ]
maintainers = [ {name = "eventyay", email = "info@eventyay.com"}, ]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Other Audience",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Environment :: Web Environment",
  "License :: OSI Approved :: Apache License, Version 2.0",
  "Programming Language :: Python :: 3.11",
  "Framework :: Django :: 5.2",
]
dependencies = [
    "aiohappyeyeballs>=2.4.4",
    "aiohttp>=3.11.11",
    "aiosignal>=1.3.2",
    "amqp>=5.3.1",
    "annotated-types>=0.7.0",
    "arabic-reshaper>=3.0.0",
    "argon2-cffi>=23.1.0",
    "argon2-cffi-bindings>=21.2.0",
    "asgiref>=3.8.1",
    "asn1crypto>=1.5.1",
    "attrs>=24.3.0",
    "autobahn>=24.4.2",
    "automat>=25.4.16",
    "babel>=2.16.0",
    "beautifulsoup4>=4.12.3",
    "billiard>=4.2.1",
    "bleach>=6.2.0",
    "cbor2>=5.6.5",
    "celery>=5.4.0",
    "certifi>=2025.4.26",
    "cffi>=1.17.1",
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.2",
    "click>=8.2.1",
    "click-didyoumean>=0.3.1",
    "click-plugins>=1.1.1",
    "click-repl>=0.3.0",
    "constantly>=23.10.4",
    "contourpy>=1.3.2",
    "coverage>=7.6.10",
    "coveralls>=4.0.1",
    "cron-descriptor>=1.4.5",
    "cryptography>=44.0.3",
    "css-inline>=0.14.6",
    "csscompressor>=0.9.5",
    "cssutils>=2.11.1",
    "cycler>=0.12.1",
    "daphne>=4.2.1",
    "defusedcsv>=2.0.0",
    "defusedxml>=0.7.1",
    "dj-static>=0.0.6",
    "django>=5.2.5",
    "django-allauth>=0.63.6",
    "django-appconf>=1.1.0",
    "django-bootstrap3>=24.3",
    "django-celery-beat>=2.7.0",
    "django-compressor>=4.5.1",
    "django-context-decorator>=1.6.1",
    "django-cors-headers>=4.6.0",
    "django-countries>=7.6.1",
    "django-csp>=3.8",
    "django-filter>=24.3",
    "django-formset-js-improved>=0.5.0.3",
    "django-formtools>=2.5.1",
    "django-hierarkey>=1.2.0",
    "django-hijack>=3.7.1",
    "django-i18nfield>=1.10.2",
    "django-jquery-js>=3.1.1",
    "django-libsass>=0.9",
    "django-localflavor>=4.0",
    "django-markup>=1.9.1",
    "django-multifactor>=0.8.2",
    "django-oauth-toolkit>=2.4.0",
    "django-otp>=1.5.4",
    "django-pdb>=0.6.2",
    "django-phonenumber-field>=7.3.0",
    "django-redis>=5.4.0",
    "django-scopes>=2.0.0",
    "django-sso>=3.0.2",
    "django-statici18n>=2.5.0",
    "django-timezone-field>=7.0",
    "djangorestframework>=3.15.2",
    "docopt>=0.6.2",
    "docutils>=0.21.2",
    "elementpath>=4.7.0",
    "emoji>=2.14.1",
    "et-xmlfile>=2.0.0",
    "eventyay-paypal @ git+https://github.com/fossasia/eventyay-tickets-paypal.git@4728e6733d971edb4db6df4ebfe12caefd371d55",
    "eventyay-stripe @ git+https://github.com/fossasia/eventyay-tickets-stripe.git@1d461de9ac7ef743cb2e6ddbace3434ba5ec6fcc",
    "execnet>=2.1.1",
    "faker>=37.3.0",
    "fakeredis>=2.26.2",
    "fido2>=1.2.0",
    "flake8>=7.1.1",
    "fonttools>=4.58.5",
    "freezegun>=1.5.1",
    "frozenlist>=1.5.0",
    "geoip2>=4.8.1",
    "gunicorn>=23.0.0",
    "hyperlink>=21.0.0",
    "icalendar>=6.3.1",
    "idna>=3.10",
    "importlib-metadata>=8.5.0",
    "incremental>=24.7.2",
    "iniconfig>=2.0.0",
    "isort>=5.13.2",
    "isoweek>=1.3.3",
    "jsonschema>=4.23.0",
    "jsonschema-specifications>=2024.10.1",
    "jwcrypto>=1.5.6",
    "kiwisolver>=1.4.8",
    "kombu>=5.5.4",
    "libsass>=0.23.0",
    "lxml>=5.4.0",
    "markdown>=3.7",
    "matplotlib>=3.10.3",
    "maxminddb>=2.6.3",
    "mccabe>=0.7.0",
    "more-itertools>=10.7.0",
    "msgpack>=1.1.1",
    "mt-940>=4.30.0",
    "multidict>=6.1.0",
    "nh3>=0.2.20",
    "numpy>=2.3.1",
    "oauthlib>=3.2.2",
    "openpyxl>=3.1.5",
    "orjson>=3.10.18",
    "packaging>=25.0",
    "pdf2image>=1.17.0",
    "pdfrw>=0.4",
    "pep8-naming>=0.14.1",
    "phonenumberslite>=8.13.52",
    "pillow>=11.0.0",
    "pluggy>=1.5.0",
    "polib>=1.2.0",
    "potypo>=0.0.2",
    "pretalx-downstream @ git+https://github.com/fossasia/eventyay-talk-downstream.git@6ad15e7bce1a621cdbf52a3d43c0f0a5e7eca664",
    "pretalx-pages @ git+https://github.com/fossasia/eventyay-talk-pages.git@a726a10459644f0bfde5019be8abf4f66b77cd27",
    "pretalx-venueless @ git+https://github.com/fossasia/eventyay-talk-video.git@8c3316a0ee7e801dc9c30b56a3c523c01209bb61",
    "pretix-pages @ git+https://github.com/fossasia/eventyay-ticket-pages.git@a7c33fd26351d409a752baa6989f24be511b5bdd",
    "pretix-venueless @ git+https://github.com/fossasia/eventyay-ticket-video.git@c7009367e02584825dc3378e5b3375a4fb613ca7",
    "prompt-toolkit>=3.0.51",
    "propcache>=0.2.1",
    "protobuf>=5.28.3",
    "psycopg[binary,pool]>=3.2.9",
    "publicsuffixlist>=1.0.2.20250529",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycodestyle>=2.12.1",
    "pycountry>=24.6.1",
    "pycparser>=2.22",
    "pycryptodome>=3.20.0",
    "pydantic>=2.10.4",
    "pydantic-core>=2.27.2",
    "pyenchant>=3.2.2",
    "pyflakes>=3.2.0",
    "pygments>=2.19.1",
    "pyjwt>=2.8.0",
    "pyopenssl>=24.3.0",
    "pyotp>=2.9.0",
    "pyparsing>=3.2.3",
    "pypdf>=5.1.0",
    "pypng>=0.20220715.0",
    "python-bidi>=0.6.3",
    "python-creole>=1.4.10",
    "python-crontab>=3.2.0",
    "python-dateutil>=2.9.0.post0",
    "python-http-client>=3.3.7",
    "python-stdnum>=1.20",
    "pytz>=2025.2",
    "pyuca>=1.2",
    "pyvat>=1.3.18",
    "pyyaml>=6.0.2",
    "qrcode>=8.2",
    "rcssmin>=1.1.2",
    "redis>=5.0.8",
    "referencing>=0.35.1",
    "regex>=2024.11.6",
    "reportlab>=4.2.5",
    "requests>=2.32.3",
    "requests-oauthlib>=2.0.0",
    "responses>=0.25.5",
    "rjsmin>=1.2.2",
    "rpds-py>=0.22.3",
    "rules>=3.5",
    "sendgrid>=6.11.0",
    "sentry-sdk>=2.19.2",
    "sepaxml>=2.6.2",
    "service-identity>=24.2.0",
    "six>=1.17.0",
    "smartypants>=2.0.1",
    "sortedcontainers>=2.4.0",
    "soupsieve>=2.7",
    "sqlparse>=0.5.3",
    "starkbank-ecdsa>=2.2.0",
    "static3>=0.7.0",
    "stripe>=11.3.0",
    "termcolor>=2.5.0",
    "text-unidecode>=1.3",
    "textile>=4.0.3",
    "tlds>=2024092600",
    "tqdm>=4.67.1",
    "twisted>=25.5.0",
    "txaio>=25.6.1",
    "typing-extensions>=4.13.2",
    "tzdata>=2025.2",
    "urllib3>=2.4.0",
    "urlman>=2.0.2",
    "vat-moss-forked>=2020.3.20.0.11.0",
    "vine>=5.1.0",
    "vobject>=0.9.9",
    "wcwidth>=0.2.13",
    "webauthn>=2.2.0",
    "webencodings>=0.5.1",
    "websockets>=15.0.1",
    "whitenoise>=6.8.2",
    "xmlschema>=3.4.3",
    "yarl>=1.18.3",
    "zipp>=3.21.0",
    "zope-interface>=7.2",
    "zxcvbn>=4.4.28",
]

[build-system]
requires = [
  'setuptools',
  'setuptools-rust',
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.4.6",
]
docs = [
    "sphinx>=8.2.3",
]
lint = [
    "ruff>=0.12.9",
]
test = [
    "pytest>=7.2.2",
    "pytest-cache>=1.0",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.9.0",
    "pytest-mock>=3.14.0",
    "pytest-rerunfailures>=14.0",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.6.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E722", "E741"]

[tool.ruff.lint.per-file-ignores]
# These files re-export models.
"**/__init__.py" = ["F401"]
# These files have long lines.
"**/migrations/*.py" = ["E501"]
